Add diagonal relations
Diagonal relations work as follows: if the expression appears on
the diagonal of a table, then the other expressions along the
diagonal are related. This is a shortcut, made possible because
table generation as implemented before this relation.

The long form is as follows: if an expression has identical
`substance` and `attribute`, or if no expression in the table has
identical `substance` and `attribute` and the `substance` has
identical `substance` and `attribute`, nested as much as
necessary, then other expressions with similarly (possibly
nested) identical `substance` and `attribute` are related.